+++ src/locales/en/meta.js
@@ -3,6 +3,10 @@
 
 export const metaMessages = {
   /* eslint-disable camelcase */
-  meta_description: "Entrepreneur from Helsinki"
+  meta_description: `An entrepreneur and doer from Etu-Töölö, Helsinki. I’m all
+  about learning something new every day and smiling as much as possible. For a
+  couple weeks I’ll still be counting the mornings left in compulsory military
+  service as Officer Candidate. In my free time I delight in volunteering in
+  NGOs and music.`
   /* eslint-enable camelcase */
 };